Create Alert context
