+++ frontend/src/App.js
@@ -1,10 +1,14 @@
-import { useState, useEffect } from "react";
+import React, { useState, useEffect } from "react";
 import CssBaseline from "@material-ui/core/CssBaseline";
 import { Switch, Route, Link, Redirect, useHistory } from "react-router-dom";
 import Layout from "./components/Layout/Layout";
 import RegistrationContainer from "./containers/RegistrationContainer/RegistrationContainer";
 import LoginContainer from "./containers/LoginContainer/LoginContainer";
 import ProfileContainer from "./containers/ProfileContainer/ProfileContainer";
+import Alert from "./components/Alert/Alert";
+
+const AlertContext = React.createContext();
+export { AlertContext };
 
 /**
  * Main function of the application
@@ -14,6 +18,13 @@ function App() {
 		isSignedIn: false,
 		accountId: null,
 	});
+
+	const [alert, setAlert] = useState({ open: false, type: "", message: "" });
+
+	function showAlert(type, message) {
+		setAlert({ open: true, type, message });
+		setTimeout(() => setAlert({ ...alert, open: false }), 7000);
+	}
 
 	let history = useHistory();
 
@@ -30,29 +41,33 @@ function App() {
 	return (
 		<>
 			<CssBaseline />
-
-			<Layout>
-				<Switch>
-					<Route exact path="/">
-						<RegistrationContainer />
-					</Route>
-					<Route path="/login">
-						<LoginContainer onLogin={onLogin} />
-					</Route>
-					<Route path="/register">
-						<RegistrationContainer />
-					</Route>
-					<Route path="/profile">
-						{signedInState.isSignedIn ? (
-							<ProfileContainer
-								accountId={signedInState.accountId}
-							/>
-						) : (
-							<Redirect to="/login" />
-						)}
-					</Route>
-				</Switch>
-			</Layout>
+			<AlertContext.Provider value={showAlert}>
+				<Layout>
+					<Switch>
+						<Route exact path="/">
+							<RegistrationContainer />
+						</Route>
+						<Route path="/login">
+							<LoginContainer onLogin={onLogin} />
+						</Route>
+						<Route path="/register">
+							<RegistrationContainer />
+						</Route>
+						<Route path="/profile">
+							{signedInState.isSignedIn ? (
+								<ProfileContainer
+									accountId={signedInState.accountId}
+								/>
+							) : (
+								<Redirect to="/login" />
+							)}
+						</Route>
+					</Switch>
+				</Layout>
+			</AlertContext.Provider>
+			{alert.open ? (
+				<Alert type={alert.type} message={alert.message} />
+			) : null}
 		</>
 	);
 }
